+++ ms-utils/math/approximation/QuotientFit.cpp
@@ -80,7 +80,7 @@ void QuotientFit::init(const double a,
     mInitFlag = false;
 
     /// - Throw a TsInitializationException exception on a singularity (divide by 0) in the allegedly valid range.
-    TS_GENERIC_IF_ERREX((mMinX < FLT_EPSILON && -FLT_EPSILON < mMaxX), TsInitializationException,
+    TS_GENERIC_IF_ERREX((mMinX < static_cast<double>(FLT_EPSILON) && -static_cast<double>(FLT_EPSILON) < mMaxX), TsInitializationException,
                         "Invalid Input Argument", "Singularity (divide by 0) in the allegedly valid range.");
 
     /// - Initialize the coefficient with validated value.
